Restructured Project File
