+++ frontend/starkbnb-web/app/components/Navbar.tsx
