+++ frontend/starkbnb-web/app/components/ui/Properties.tsx
